Remove the recommendation to use random bytes for nonces
My reading of randombytes_buf() is that it generates a byte buffer from a
source of random data that is likely to be uniformly random.

Although it's highly likely that crypto_box_NONCEBYTES of data will be unique
for most applicaiton invokations, there doesn't appear to be any kind of
gaurantee; and how should the implementor reason about re-keying? When they've
sent 2 ^ (crypto_box_NONCEBYTES * 8) bytes, or before?

Would things be simpler and more clear with partitionable sequences?