Fixing crash issue on Linux x86
On Linux x86, SIZE_MAX is 2^32, but an unsigned long long int is 2^64. In randombytes.h, the struct randombytes_implementation defines buf with the second parameter being a size_t (2^32), but the randombytes() function is defined using a long long unsigned int (2^64). This causes the library to try and read 64 bits of data for the buffer length when there are only 32 bits available, which causes a failed assertion.