Slightly revamp dist-build/ios.sh - Compile arm32 and arm64 targets separately - Compile for the iOS simulator
The resulting fat library include arm32, arm64, 32-bit simulator and
64-bit OSX targets.

This is a bit better than what we previously had regarding the simulator,
but still not optimal, as the 64-bit simulator version should also be
included. However, a fat library cannot include both the 64-bit simulator
version and a version for OSX since the architecture is the same.

The proper way to address this is probably to generate two libraries.
This requires some prior changes to Swift-Sodium.