Make the Salsa20 RNG build and work on CloudABI.
Though CloudABI is a POSIX-like runtime environment, there is no
getpid(). Processes cannot observe their process ID.

This code only needs to know the PID to reset the randomizer state in
case a fork() is performed. Simply use pthread_atfork() to register a
handler that tears down the RNG when forking.