add crypto_scalarmult_ed25519_noclamp
`crypto_scalarmult_ed25519(q, n, p)` clamps the scalar `n` to a suitable
value for finding the point on the curve. This is limiting in scenarios
where the scalar multiplication should be done over unclamped scalars
such as field elements arising out of field arithmetic. An example of
such a use case is the Schnorr NIZK protocol from RFC8235.