Modify configure.ac to actually respect --disable-asm.
The --disable-asm option is supposed to disable all assembly or
platform-specific optimizations, but actually only managed to disable
x86_64-specific assembly code. This adds a WITH_ASM autoconf variable --
and an automake definition of the same name -- set to 0 or 1 reflecting
whether ASM is enabled or disabled for the build.

As one might expect, the existing HAVE_AMD64_ASM and HAVE_AVX_ASM flags
will only be defined if HAVE_ASM is also defined.