Fix: make only a single call to xcodebuild -create-xcframework
xcodebuild -create-xcframework writes manifest metadata to the root of the XCFramework in Info.plist. Although calling it repeatedly (instead of making a single call) does add the new subfolders to the bundle without wiping the previous ones, it _doesn't_ merge the old + new manifest data together; instead, each subsequent call clobbers any existing Info.plist. As a result, the only way to use it is to make a single call, and include all of the library+header pairs at once.