Add sodium_finalize()
`sodium_crit_enter()` lazily initializes a critical section on Windows
but that critical section is never deleted (`DeleteCriticalSection()`).
This is not an issue per se, since the critical section is released on
process termination, but Application Verifier complains about that
("Unloading DLL containing an active critical section.")

There is no way yet to delete this critical section, since it is held
in a file static variable, and there is no API to delete it.  Thus,
I suggest to introduce `sodium_finalize()` which can be called by
clients to delete the critical section on process termination.

The new API should be available on all supported platforms, although
I don't know whether similar issues exists elsewhere.  For now, it does
nothing on other platforms than Windows.