Use assignment instead of memcpy to copy the value of a pointer
Some static code analyzers report warnings when `sizeof` is used on a
pointer type, as this can be the sign of a missing dereference operation.
For example clang's static analyzer has an alpha checker to report such
issues, `alpha.core.SizeofPtr`.

In function `allocate_memory`, `sizeof` is used on a pointer:

    void * base;
    block *memory;

    /* ... */

    memcpy(&memory, &base, sizeof memory);

This triggers a warning, and this call to `memcpy` can safely be
replaced by a usual assignment (the variables are not special in a way,
they are aligned using the default alignment, etc.):

    memory = base;

Silences a warning from some static analyzers by changing this.