Add proper CET support
Use cet.h if available and use _CET_ENDBR to make all the functions
IBT-safe.  This is not directly testable yet, but the instruction is in
the nop space for processors that don't support it, so it's a harmless
addition.

cet.h also includes the SHSTK and IBT GNU properties in ELF, so remove
the manually added ones.