Allow other zig projects to use libsodium as a dependency (#1300)
* fix(build.zig): derive cwd from builder

Using fs.cwd() only resolves the libsodium directory when building from
within libsodium. Deriving the absolute path to the libsodium directory
from the builder allows the build script to be invoked from other
projects.

* feat(build.zig): add options for building libsodium

Added the options to build either the static, shared, and/or tests. The
tests depend on the static library, and therefore imply the static
library.

* feat(build.zig): include header files